requirements.txt
pika==1.2.0
pyyaml==5.3.1
python-packer==0.1.2